promote version-specific functions to a proper helper file
