+++ src/polyglot-notebooks-vscode-common/src/extension.ts
@@ -25,7 +25,7 @@ import { OutputChannelAdapter } from './OutputChannelAdapter';
 
 import * as notebookControllers from './notebookControllers';
 import * as notebookSerializers from './notebookSerializers';
-import * as versionSpecificFunctions from '../versionSpecificFunctions';
+import * as vscodeNotebookManagement from './vscodeNotebookManagement';
 import { ErrorOutputCreator, InteractiveClient } from './interactiveClient';
 
 import * as vscodeUtilities from './vscodeUtilities';
@@ -163,7 +163,7 @@ export async function activate(context: vscode.ExtensionContext) {
                 const password = requestInput.isPassword;
 
                 let value;
-                let customInputRequest = await versionSpecificFunctions.handleCustomInputRequest(prompt, requestInput.inputTypeHint, password);
+                let customInputRequest = await vscodeNotebookManagement.handleCustomInputRequest(prompt, requestInput.inputTypeHint, password);
                 if (customInputRequest.handled) {
                     value = customInputRequest.result;
                 } else {
@@ -217,7 +217,7 @@ export async function activate(context: vscode.ExtensionContext) {
                 };
                 const rawCellMetadata = metadataUtilities.getRawNotebookCellMetadataFromNotebookCellMetadata(notebookCellMetadata);
                 newCell.metadata = rawCellMetadata;
-                const succeeded = await versionSpecificFunctions.replaceNotebookCells(notebookDocument.uri, range, [newCell]);
+                const succeeded = await vscodeNotebookManagement.replaceNotebookCells(notebookDocument.uri, range, [newCell]);
 
                 if (!succeeded) {
                     throw new Error(`Unable to add cell to notebook '${notebookUri.toString()}'.`);
