+++ src/polyglot-notebooks-vscode-common/src/languageProvider.ts
@@ -9,7 +9,7 @@ import { notebookCellChanged } from './interactiveNotebook';
 import { convertToRange, getCellKernelName, toVsCodeDiagnostic } from './vscodeUtilities';
 import { getDiagnosticCollection } from './diagnostics';
 import { provideSignatureHelp } from './languageServices/signatureHelp';
-import * as versionSpecificFunctions from '../versionSpecificFunctions';
+import * as vscodeNotebookManagement from './vscodeNotebookManagement';
 import * as constants from './constants';
 
 function getNotebookDcoumentFromCellDocument(cellDocument: vscode.TextDocument): vscode.NotebookDocument | undefined {
@@ -168,7 +168,7 @@ export async function registerLanguageProviders(clientMapper: ClientMapper, lang
     disposables.push(vscode.languages.registerSignatureHelpProvider(languages, new SignatureHelpProvider(clientMapper, languageServiceDelay), ...SignatureHelpProvider.triggerCharacters));
     disposables.push(vscode.workspace.onDidChangeTextDocument(e => {
         if (e.document.languageId === constants.CellLanguageIdentifier && vscode.window.activeNotebookEditor) {
-            const notebookDocument = versionSpecificFunctions.getNotebookDocumentFromEditor(vscode.window.activeNotebookEditor);
+            const notebookDocument = vscodeNotebookManagement.getNotebookDocumentFromEditor(vscode.window.activeNotebookEditor);
             const cells = notebookDocument.getCells();
             const cell = cells?.find(cell => cell.document === e.document);
             if (cell) {
