+++ src/polyglot-notebooks-vscode-common/src/notebookControllers.ts
@@ -15,7 +15,7 @@ import { isKernelEventEnvelope, KernelCommandOrEventEnvelope } from './polyglot-
 import * as rxjs from 'rxjs';
 import * as metadataUtilities from './metadataUtilities';
 import * as constants from './constants';
-import * as versionSpecificFunctions from '../versionSpecificFunctions';
+import * as vscodeNotebookManagement from './vscodeNotebookManagement';
 import * as semanticTokens from './documentSemanticTokenProvider';
 import { ServiceCollection } from './serviceCollection';
 
@@ -138,7 +138,7 @@ export class DotNetNotebookKernel {
                     this.config.clientMapper.getOrAddClient(notebookUri).then(() => {
                         const kernelInfoProduced = (<contracts.KernelEventEnvelope[]>(e.message.kernelInfoProduced)).map(e => <contracts.KernelInfoProduced>e.event);
                         const hostUri = e.message.hostUri;
-                        versionSpecificFunctions.hashBangConnect(this.config.clientMapper, hostUri, kernelInfoProduced, this.uriMessageHandlerMap, (arg) => controller.postMessage(arg), notebookUri);
+                        vscodeNotebookManagement.hashBangConnect(this.config.clientMapper, hostUri, kernelInfoProduced, this.uriMessageHandlerMap, (arg) => controller.postMessage(arg), notebookUri);
                     });
                     break;
             }
@@ -311,7 +311,7 @@ async function updateKernelInfoMetadata(client: InteractiveClient, document: vsc
                 const existingRawNotebookDocumentMetadata = document.metadata;
                 const updatedRawNotebookDocumentMetadata = metadataUtilities.getRawNotebookDocumentMetadataFromNotebookDocumentMetadata(notebookMetadata, isIpynb);
                 const newRawNotebookDocumentMetadata = metadataUtilities.mergeRawMetadata(existingRawNotebookDocumentMetadata, updatedRawNotebookDocumentMetadata);
-                await versionSpecificFunctions.replaceNotebookMetadata(document.uri, newRawNotebookDocumentMetadata);
+                await vscodeNotebookManagement.replaceNotebookMetadata(document.uri, newRawNotebookDocumentMetadata);
             }
         }
     });
@@ -320,7 +320,7 @@ async function updateKernelInfoMetadata(client: InteractiveClient, document: vsc
     const kernelNotebokMetadata = metadataUtilities.getNotebookDocumentMetadataFromCompositeKernel(client.kernel);
     const mergedMetadata = metadataUtilities.mergeNotebookDocumentMetadata(notebookDocumentMetadata, kernelNotebokMetadata);
     const rawNotebookDocumentMetadata = metadataUtilities.getRawNotebookDocumentMetadataFromNotebookDocumentMetadata(mergedMetadata, isIpynb);
-    await versionSpecificFunctions.replaceNotebookMetadata(document.uri, rawNotebookDocumentMetadata);
+    await vscodeNotebookManagement.replaceNotebookMetadata(document.uri, rawNotebookDocumentMetadata);
 }
 
 async function updateCellLanguagesAndKernels(document: vscode.NotebookDocument): Promise<void> {
@@ -372,5 +372,5 @@ function generateVsCodeNotebookCellOutputItem(data: Uint8Array, mime: string, st
 async function updateDocumentKernelspecMetadata(document: vscode.NotebookDocument): Promise<void> {
     const documentMetadata = metadataUtilities.getNotebookDocumentMetadataFromNotebookDocument(document);
     const newMetadata = metadataUtilities.createNewIpynbMetadataWithNotebookDocumentMetadata(document.metadata, documentMetadata);
-    await versionSpecificFunctions.replaceNotebookMetadata(document.uri, newMetadata);
+    await vscodeNotebookManagement.replaceNotebookMetadata(document.uri, newMetadata);
 }
