+++ src/polyglot-notebooks-vscode-common/src/variableExplorer.ts
@@ -6,7 +6,7 @@ import * as vscode from 'vscode';
 import { ClientMapper } from './clientMapper';
 import * as contracts from './polyglot-notebooks/contracts';
 import * as utilities from './utilities';
-import * as versionSpecificFunctions from '../versionSpecificFunctions';
+import * as vscodeNotebookManagement from './vscodeNotebookManagement';
 import { Disposable } from './polyglot-notebooks/disposables';
 import { isKernelEventEnvelope } from './polyglot-notebooks';
 import * as kernelSelectorUtilities from './kernelSelectorUtilities';
@@ -21,7 +21,7 @@ export function registerVariableExplorer(context: vscode.ExtensionContext, clien
     context.subscriptions.push(vscode.commands.registerCommand('polyglot-notebook.shareValueWith', async (variableInfo: VariableInfo | undefined) => {
         const activeNotebookEditor = vscode.window.activeNotebookEditor;
         if (variableInfo && activeNotebookEditor) {
-            const notebookDocument = versionSpecificFunctions.getNotebookDocumentFromEditor(activeNotebookEditor);
+            const notebookDocument = vscodeNotebookManagement.getNotebookDocumentFromEditor(activeNotebookEditor);
             const client = await clientMapper.tryGetClient(notebookDocument.uri);
             if (client) {
                 const kernelSelectorOptions = kernelSelectorUtilities.getKernelSelectorOptions(client.kernel, notebookDocument, contracts.SendValueType);
